+++ scalr/provider.go
@@ -68,13 +68,15 @@ func Provider() terraform.ResourceProvider {
 			"scalr_current_run":   dataSourceScalrCurrentRun(),
 			"scalr_endpoint":      dataSourceScalrEndpoint(),
 			"scalr_webhook":       dataSourceScalrWebhook(),
+			"scalr_environment":   dataSourceScalrEnvironment(),
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
-			"scalr_workspace": resourceScalrWorkspace(),
-			"scalr_variable":  resourceScalrVariable(),
-			"scalr_endpoint":  resourceScalrEndpoint(),
-			"scalr_webhook":   resourceScalrWebhook(),
+			"scalr_workspace":   resourceScalrWorkspace(),
+			"scalr_variable":    resourceScalrVariable(),
+			"scalr_endpoint":    resourceScalrEndpoint(),
+			"scalr_webhook":     resourceScalrWebhook(),
+			"scalr_environment": resourceScalrEnvironment(),
 		},
 
 		ConfigureFunc: providerConfigure,